Mutate the object from the getBooks HTTP GET call from the action function
