+++ src/redux/configureStore.js
@@ -1,4 +1,7 @@
-import { createStore, combineReducers, applyMiddleware } from 'redux';
+/* eslint-disable no-underscore-dangle */
+import {
+  createStore, combineReducers, applyMiddleware, compose,
+} from 'redux';
 import thunk from 'redux-thunk';
 import booksReducer from './books/books';
 
@@ -8,7 +11,10 @@ const reducer = combineReducers({
 
 const store = createStore(
   reducer,
-  applyMiddleware(thunk),
+  compose(
+    applyMiddleware(thunk),
+    window.__REDUX_DEVTOOLS_EXTENSION__ && window.__REDUX_DEVTOOLS_EXTENSION__(),
+  ),
 );
 
 export default store;
